+++ src/test/resources/testing.xml
@@ -1,11 +1,16 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 <suite name="Use tests automation" parallel="methods" thread-count="5">
 
+    <groups>
+        <run>
+            <!--<include name="GettersTest"/>-->
+            <include name="All"/>
+        </run>
+    </groups>
     <test name="Operations smoke">
         <classes>
-
-            <class name="Tests"/>
+            <class name="TestWoman"/>
+            <class name="TestMan"/>
         </classes>
-
     </test>
 </suite>
